refactor : (Gym) Facade 적용, Service 추상화
